+++ lib/zabbix/request.rb
@@ -1,8 +1,8 @@
+# frozen_string_literal: true
 
 module Zabbix
   # Deals with requests
   module Request
-
     # JSON-RPC is a stateless, light-weight remote procedure call (RPC) protocol. Primarily this
     # specification defines several data structures and the rules around their processing. It is
     # transport agnostic in that the concepts can be used within the same process, over sockets, over
@@ -16,21 +16,20 @@ module Zabbix
 
       def rpc_call(method, params = nil)
         options = {
-          "jsonrpc": "2.0",
-          "method": method,
-          "params": {
-            "output": "extend"
+          'jsonrpc': '2.0',
+          'method': method,
+          'params': {
+            'output': 'extend'
           },
-          "id": @@id += 1
+          'id': @@id += 1
         }
 
         options[:params] = params if params
-        result = post( ZABBIX_ENDPOINT, options )
+        result = post(ZABBIX_ENDPOINT, options)
         raise RPCError, result.body['error'] if result.body['error']
 
         WrAPI::Request::Entity.create(result.body['result'])
       end
-
     end
   end
 end