+++ lib/zabbix_api_gem.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 
 # place holder to use unique gem name
 require File.expand_path('zabbix', __dir__)