+++ test/auth_test.rb
@@ -1,7 +1,9 @@
-require 'logger'
-require "test_helper"
+# frozen_string_literal: true
 
-AUTH_LOGGER = "auth_test.log"
+require 'logger'
+require 'test_helper'
+
+AUTH_LOGGER = 'auth_test.log'
 File.delete(AUTH_LOGGER) if File.exist?(AUTH_LOGGER)
 
 describe 'auth' do
@@ -9,16 +11,16 @@ describe 'auth' do
     Zabbix.reset
     Zabbix.logger = Logger.new(AUTH_LOGGER)
   end
-  it "#0 check required params" do
+  it '#0 check required params' do
     c = Zabbix.client
     # missing endpoint
     assert_raises Zabbix::ConfigurationError do
       c.login
     end
   end
-  it "#1 check required params" do
+  it '#1 check required params' do
     Zabbix.configure do |config|
-      config.endpoint = ENV["ZABBIX_API_HOST"]
+      config.endpoint = ENV['ZABBIX_API_HOST']
     end
     c = Zabbix.client
     # missing access_token
@@ -26,9 +28,9 @@ describe 'auth' do
       c.login
     end
   end
-  it "#2 wrong credentials" do
+  it '#2 wrong credentials' do
     Zabbix.configure do |config|
-      config.endpoint = ENV["ZABBIX_API_HOST"]
+      config.endpoint = ENV['ZABBIX_API_HOST']
       config.access_token = 'api-key-token'
     end
     c = Zabbix.client
@@ -36,13 +38,13 @@ describe 'auth' do
       c.login
     end
   end
-  it "#3 logged in" do
+  it '#3 logged in' do
     Zabbix.configure do |config|
-      config.endpoint = ENV["ZABBIX_API_HOST"]
-      config.access_token = ENV["ZABBIX_API_KEY"]
+      config.endpoint = ENV['ZABBIX_API_HOST']
+      config.access_token = ENV['ZABBIX_API_KEY']
     end
     c = Zabbix.client
 
-    refute_empty c.login, ".login"
+    refute_empty c.login, '.login'
   end
 end