+++ test/test_helper.rb
@@ -1,9 +1,10 @@
 # frozen_string_literal: true
+
 require 'simplecov'
 SimpleCov.start
 
-$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
-require 'Dotenv'
+$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
+require 'dotenv'
 require 'zabbix_api_gem'
 require 'minitest/autorun'
 require 'minitest/spec'