+++ CHANGELOG.md
@@ -1,20 +1,25 @@
-## [Unreleased]
+# Changelog
 
 ## [0.1.0] - 2024-02-09
+
 - Initial release
 
 ## [0.2.0] - 2024-02-12
+
 - Include events and method to get object by id
 
 ## [0.3.0] - 2024-02-20
+
 - ConfigurationError raised if not configured properly
 
 ## [0.3.1] - 2024-02-22
+
 - Fix problem with entity creation
 
 ## [0.3.2] - 2024-03-21
+
 - Refactory extract method
 
 ## [0.3.3] - 2024-03-21
-- Fix issues with extract method
 
+- Fix issues with extract method